chore: adjust preference trees
- Move `Schemas`, `Appearance`(`Themes` and `Colors`) to top tree
- Refactor some preference items' titles
- Hide the options menu in sub settings page
