+++ app/src/main/java/com/osfans/trime/settings/fragments/AppearanceFragment.kt
@@ -1,0 +1,38 @@
+package com.osfans.trime.settings.fragments
+
+import android.os.Bundle
+import android.view.Menu
+import androidx.core.view.forEach
+import androidx.core.view.forEachIndexed
+import androidx.preference.Preference
+import androidx.preference.PreferenceFragmentCompat
+import com.osfans.trime.R
+import com.osfans.trime.settings.components.ColorPickerDialog
+import com.osfans.trime.settings.components.ThemePickerDialog
+
+class AppearanceFragment: PreferenceFragmentCompat() {
+    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
+        addPreferencesFromResource(R.xml.appearance_preference)
+
+        setHasOptionsMenu(true)
+    }
+
+    override fun onPrepareOptionsMenu(menu: Menu) {
+        menu.forEach { item -> item.isVisible = false}
+        super.onPrepareOptionsMenu(menu)
+    }
+
+    override fun onPreferenceTreeClick(preference: Preference?): Boolean {
+        return when (preference?.key) {
+            "pref_themes" -> {
+                ThemePickerDialog(requireContext()).show()
+                true
+            }
+            "pref_colors" -> {
+                ColorPickerDialog(requireContext()).show()
+                true
+            }
+            else -> super.onPreferenceTreeClick(preference)
+        }
+    }
+}
