+++ app/src/main/java/com/osfans/trime/settings/fragments/KeyboardFragment.kt
@@ -2,6 +2,8 @@ package com.osfans.trime.settings.fragments
 
 import android.content.SharedPreferences
 import android.os.Bundle
+import android.view.Menu
+import androidx.core.view.forEach
 import androidx.preference.Preference
 import androidx.preference.PreferenceFragmentCompat
 import com.osfans.trime.R
@@ -14,20 +16,13 @@ class KeyboardFragment: PreferenceFragmentCompat(),
     SharedPreferences.OnSharedPreferenceChangeListener {
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         addPreferencesFromResource(R.xml.keyboard_preference)
+
+        setHasOptionsMenu(true)
     }
 
-    override fun onPreferenceTreeClick(preference: Preference?): Boolean {
-        return when (preference?.key) {
-            "pref_themes" -> {
-                ThemePickerDialog(requireContext()).show()
-                true
-            }
-            "pref_colors" -> {
-                ColorPickerDialog(requireContext()).show()
-                true
-            }
-            else -> super.onPreferenceTreeClick(preference)
-        }
+    override fun onPrepareOptionsMenu(menu: Menu) {
+        menu.forEach { item -> item.isVisible = false}
+        super.onPrepareOptionsMenu(menu)
     }
 
     override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences?, key: String?) {
