+++ app/src/main/res/xml/appearance_preference.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <Preference android:key="pref_themes"
+        app:iconSpaceReserved="false"
+        android:title="@string/pref_themes"
+        android:summary="@string/pref_themes_summary" />
+    <Preference android:key="pref_colors"
+        app:iconSpaceReserved="false"
+        android:title="@string/pref_colors"
+        android:summary="@string/pref_colors_summary" />
+</PreferenceScreen>
